feat(metronome): implement entering bpm using the keyboard
