+++ lib/music_player/slowdowner/slowdowner_card.dart
@@ -6,6 +6,7 @@ import 'package:flutter/services.dart';
 import 'package:musbx/music_player/card_header.dart';
 import 'package:musbx/music_player/music_player.dart';
 import 'package:musbx/music_player/slowdowner/circular_slider/circular_slider.dart';
+import 'package:musbx/widgets.dart';
 
 class SlowdownerCard extends StatelessWidget {
   /// Card with sliders for changing the pitch and speed of [MusicPlayer].
@@ -88,7 +89,7 @@ class SlowdownerCard extends StatelessWidget {
             builder: (context, constraints) {
               return ConstrainedBox(
                 constraints: BoxConstraints(maxWidth: constraints.maxWidth / 2),
-                child: NumberField(
+                child: NumberField<double>(
                   value: double.parse(pitch.toStringAsFixed(1)),
                   min: -12.0,
                   max: 12.0,
@@ -147,7 +148,7 @@ class SlowdownerCard extends StatelessWidget {
             builder: (context, constraints) {
               return ConstrainedBox(
                 constraints: BoxConstraints(maxWidth: constraints.maxWidth / 2),
-                child: NumberField(
+                child: NumberField<double>(
                   value: double.parse(speed.toStringAsFixed(2)),
                   min: 0.5,
                   max: 2.0,
@@ -194,92 +195,6 @@ class SlowdownerCard extends StatelessWidget {
                 },
           icon: const Icon(Icons.refresh),
         ),
-      ),
-    );
-  }
-}
-
-class NumberField extends StatelessWidget {
-  NumberField({
-    super.key,
-    this.value,
-    this.min,
-    this.max,
-    required this.onSubmitted,
-    this.prefixWithSign = true,
-    this.style,
-    this.inputFormatters,
-  });
-
-  final double? value;
-  final double? min;
-  final double? max;
-
-  /// If `true`, automatically prefixes the value with the correct sign.
-  ///
-  /// When the user starts editing the value, the prefix is removed so that the sign can be edited.
-  final bool prefixWithSign;
-  final TextStyle? style;
-  final List<TextInputFormatter>? inputFormatters;
-
-  final void Function(double value)? onSubmitted;
-
-  late final TextEditingController controller = TextEditingController(
-    text: prefixWithSign ? value?.abs().toString() : value?.toString(),
-  );
-
-  /// Optional text prefix to place on the line before the input.
-  late final ValueNotifier<String?> prefixText =
-      ValueNotifier(!prefixWithSign || value == null
-          ? null
-          : value! >= 0
-              ? "+"
-              : "-");
-
-  @override
-  Widget build(BuildContext context) {
-    return ListenableBuilder(
-      listenable: prefixText,
-      builder: (context, child) => TextField(
-        controller: controller,
-        decoration: InputDecoration(
-          prefixText: prefixText.value,
-          border: const UnderlineInputBorder(),
-          contentPadding: const EdgeInsets.symmetric(vertical: 4.0),
-        ),
-        style: style,
-        textAlign: TextAlign.center,
-        textAlignVertical: TextAlignVertical.center,
-        keyboardType: TextInputType.number,
-        inputFormatters: [
-          FilteringTextInputFormatter.allow(RegExp(
-              r"^(-|-?\d+(?:\.\d*)?)$")), // Allow positive and negative decimal numbers, as well as just a minus sign
-          ...?inputFormatters,
-        ],
-        maxLines: 1,
-        enabled: onSubmitted != null,
-        onChanged: (value) {
-          /// Remove prefix and instead add the minus sign to the text, so that it can be edited.
-          if (prefixText.value == "-" && !value.startsWith("-")) {
-            controller.text = "-${controller.text}";
-          }
-          prefixText.value = null;
-        },
-        onSubmitted: (value) {
-          double? parsed = double.tryParse(value);
-          if (parsed == null) {
-            // Reset to the actual value
-            if (this.value != null) controller.text = this.value.toString();
-            return;
-          }
-
-          double clamped = parsed.clamp(
-            min ?? double.negativeInfinity,
-            max ?? double.infinity,
-          );
-          controller.text = clamped.toString();
-          onSubmitted?.call(clamped);
-        },
       ),
     );
   }
